Distinct contexts for 'in' and 'annotation' parts
Atkey's rule specifies a context extended with z (Local ii) in the typing
judgement of the `type annotation` part, but we used the same context,
extended with x (Local ii) and y (Local ii+1), which is supposed to
be used only in the typing judgement of the `in` part of the eliminator.

Along with this, we improved the readability of the typing rules.