Fix name clashes in pretty-printer
The original pretty-printer produced wrong results, because it was able
to pick a name for a bound variable that was already used for a free
variable in the given term. This lead to a term being printed that had
different semantics from the actual term. For example, term:

  let 0 fn = \a. x : (0 _ : U) -> X

would print as

  fn = 0 \x. x : (0 x : U) -> X.

Now, the State monad collects all the free variables occurring in
the term and then the Reader monad provides a list of names that can
be used to name a bound variable.